+++ pubsub/subscriber.go
@@ -16,12 +16,14 @@ type Handler interface {
 }
 
 // NewSubscriber creates an instance of the provided adapter
-func NewSubscriber(node Handler, adapter string, redis *RedisConfig, http *HTTPConfig) (Subscriber, error) {
+func NewSubscriber(node Handler, adapter string, redis *RedisConfig, http *HTTPConfig, nats *NATSConfig) (Subscriber, error) {
 	switch adapter {
 	case "redis":
 		return NewRedisSubscriber(node, redis), nil
 	case "http":
 		return NewHTTPSubscriber(node, http), nil
+	case "nats":
+		return NewNATSSubscriber(node, nats), nil
 	}
 
 	return nil, fmt.Errorf("Unknown adapter type: %s", adapter)